requirements_clean.txt
# Core web framework and server
Flask==3.1.1
flask-cors==6.0.0
gunicorn

# AI and LangChain dependencies
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-google-genai==2.1.4
langchain-text-splitters==0.3.8
langgraph==0.4.7
langgraph-checkpoint==2.0.26
langgraph-prebuilt==0.2.1

# Google AI integration
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0

# Search and utilities
duckduckgo_search==8.0.2
python-dotenv==1.1.0

# Core Python dependencies
requests==2.32.3
pydantic==2.11.5
pydantic_core==2.33.2
PyYAML==6.0.2
